# public/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "macvendor API",
    "version": "1.0.0",
    "description": "Source-aware MAC address block assignment lookup. Synthetic examples only; an assignment is not proof of a physical device manufacturer."
  },
  "servers": [
    { "url": "https://macvendor.io", "description": "Maintained public service" }
  ],
  "paths": {
...
    "/v1/assignments/{registry}/{prefix}": {
      "get": {
        "operationId": "getAssignment",
        "parameters": [
          { "name": "registry", "in": "path", "required": true, "schema": { "type": "string", "enum": ["ma-l", "ma-m", "ma-s", "iab", "cid"] }, "example": "ma-l" },
          { "name": "prefix", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^[0-9A-F]+-(24|28|36)$" }, "example": "02AABB-24" },
          { "name": "include", "in": "query", "schema": { "type": "string", "const": "evidence" } }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/AssignmentSuccess" },
          "308": { "$ref": "#/components/responses/CanonicalRedirect" },
          "400": { "$ref": "#/components/responses/Problem" },
          "404": { "$ref": "#/components/responses/Problem" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
...
  },
  "components": {
    "headers": {
      "RequestId": { "description": "Opaque request identifier", "schema": { "type": "string" } },
      "ApiVersion": { "description": "Stable public API contract version", "schema": { "type": "string", "const": "v1" } },
      "AppVersion": { "description": "Deployed application SemVer", "schema": { "type": "string", "pattern": "^[0-9]+\\.[0-9]+\\.[0-9]+$" } },
      "CacheControl": { "description": "Representation cache policy", "schema": { "type": "string" } },
...
      "Location": { "description": "Canonical URL", "schema": { "type": "string", "format": "uri" } },
      "RetryAfter": { "description": "Seconds before retry", "schema": { "type": "integer", "minimum": 1 } }
    },
    "responses": {
...
      "AssignmentSuccess": {
        "description": "Exact active assignment",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" }, "X-API-Version": { "$ref": "#/components/headers/ApiVersion" }, "X-App-Version": { "$ref": "#/components/headers/AppVersion" }, "Cache-Control": { "$ref": "#/components/headers/CacheControl" } },
        "content": { "application/json": { "schema": { "$ref": "/schemas/public-api-v1.schema.json#/$defs/AssignmentResponse" } } }
      },
...
      "CanonicalRedirect": {
        "description": "Permanent redirect to the canonical path while preserving the query",
        "headers": {
          "X-Request-Id": { "$ref": "#/components/headers/RequestId" },
          "X-API-Version": { "$ref": "#/components/headers/ApiVersion" },
          "X-App-Version": { "$ref": "#/components/headers/AppVersion" },
          "Cache-Control": { "$ref": "#/components/headers/CacheControl" },
          "Location": { "$ref": "#/components/headers/Location" }
        }
      },
      "Problem": {
        "description": "RFC 9457 problem response",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" }, "X-API-Version": { "$ref": "#/components/headers/ApiVersion" }, "X-App-Version": { "$ref": "#/components/headers/AppVersion" }, "Cache-Control": { "$ref": "#/components/headers/CacheControl" } },
        "content": { "application/problem+json": { "schema": { "$ref": "/schemas/public-api-v1.schema.json#/$defs/Problem" } } }
      },
      "RateLimited": {
        "description": "Rate limit exceeded",
        "headers": {
          "X-Request-Id": { "$ref": "#/components/headers/RequestId" },
          "X-API-Version": { "$ref": "#/components/headers/ApiVersion" },
          "X-App-Version": { "$ref": "#/components/headers/AppVersion" },
          "Cache-Control": { "$ref": "#/components/headers/CacheControl" },
          "Retry-After": { "$ref": "#/components/headers/RetryAfter" }
        },
        "content": { "application/problem+json": { "schema": { "$ref": "/schemas/public-api-v1.schema.json#/$defs/Problem" } } }
      }
    }
  }
}